Update too much stuff.
Unfortunately, it has become a too difficult to merge in the newest content type changes from ding2tal.
This is mainly due to field type changes and maybe other changes that features doesn't handle well.
We have to work on avoiding data loss when pulling in changes to the content types.